requirements.txt
aiogram>=3.4.0,<4.0
asyncpg>=0.29.0
redis[hiredis]>=5.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
arq>=0.25.0
aiohttp>=3.9.0
cryptography>=42.0.0
PyJWT>=2.8.0
httpx>=0.27.0
tenacity>=8.2.3
pydantic-settings>=2.1.0
sentence-transformers>=2.2.2
pymupdf>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
pgvector>=0.3.0
numpy>=1.26.0
